app/requirements.txt
kubernetes
google.generativeai
requests